+++ regolith/htmlbuilder.py
@@ -1,16 +1,13 @@
 """Builder for websites."""
 import os
 import shutil
-from copy import deepcopy
 from itertools import groupby
 
-from jinja2 import Environment, FileSystemLoader
 
-from regolith.latex import make_bibtex_file
 from regolith.builder_base import BuilderBase
 
 from regolith.tools import all_docs_from_collection, date_to_rfc822, \
-    rfc822now, gets
+    rfc822now, gets, filter_publications, filter_projects, make_bibtex_file
 from regolith.sorters import doc_date_key, ene_date_key, category_val, \
     level_val, id_key, date_key, position_key
 
@@ -66,12 +63,15 @@ class HtmlBuilder(BuilderBase):
         os.makedirs(former_peeps_dir, exist_ok=True)
         for p in self.gtx['people']:
             names = frozenset(p.get('aka', []) + [p['name']])
-            pubs = self.filter_publications(names, reverse=True)
+            pubs = filter_publications(
+                all_docs_from_collection(rc.client, 'citations'),
+                names, reverse=True, bold=False)
             bibfile = make_bibtex_file(pubs, pid=p['_id'],
                                        person_dir=peeps_dir)
             ene = p.get('employment', []) + p.get('education', [])
             ene.sort(key=ene_date_key, reverse=True)
-            projs = self.filter_projects(names)
+            projs = filter_projects(
+                all_docs_from_collection(rc.client, 'projects'), names)
             self.render('person.html',
                         os.path.join('people', p['_id'] + '.html'), p=p,
                         title=p.get('name', ''), pubs=pubs, names=names,
@@ -82,32 +82,6 @@ class HtmlBuilder(BuilderBase):
 
         self.render('former.html', os.path.join('former', 'index.html'),
                     title='Former Members')
-
-    def filter_publications(self, authors, reverse=False):
-        rc = self.rc
-        pubs = []
-        for pub in all_docs_from_collection(rc.client, 'citations'):
-            s = set(pub.get('author', []))
-            # XXX: If editor but not author, but breaks multiple places
-            # s |= set(pub.get('editor', []))
-            if len(s & authors) == 0:
-                continue
-            pubs.append(deepcopy(pub))
-        pubs.sort(key=doc_date_key, reverse=reverse)
-        return pubs
-
-    def filter_projects(self, authors, reverse=False):
-        rc = self.rc
-        projs = []
-        for proj in all_docs_from_collection(rc.client, 'projects'):
-            team_names = set(gets(proj['team'], 'name'))
-            if len(team_names & authors) == 0:
-                continue
-            proj = dict(proj)
-            proj['team'] = [x for x in proj['team'] if x['name'] in authors]
-            projs.append(proj)
-        projs.sort(key=id_key, reverse=reverse)
-        return projs
 
     def projects(self):
         rc = self.rc
@@ -131,7 +105,6 @@ class HtmlBuilder(BuilderBase):
         self.render('rss.xml', os.path.join('blog', 'rss.xml'), items=posts)
 
     def jobs(self):
-        rc = self.rc
         jobs_dir = os.path.join(self.bldir, 'jobs')
         os.makedirs(jobs_dir, exist_ok=True)
         for job in self.gtx['jobs']:
@@ -142,7 +115,6 @@ class HtmlBuilder(BuilderBase):
                     title='Jobs')
 
     def abstracts(self):
-        rc = self.rc
         abs_dir = os.path.join(self.bldir, 'abstracts')
         os.makedirs(abs_dir, exist_ok=True)
         for ab in self.gtx['abstracts']:
