+++ regolith/tools.py
@@ -5,7 +5,20 @@ import re
 import sys
 import platform
 import email.utils
+from copy import deepcopy
+
 from datetime import datetime
+
+from regolith.sorters import doc_date_key, id_key, ene_date_key
+
+try:
+    from bibtexparser.bwriter import BibTexWriter
+    from bibtexparser.bibdatabase import BibDatabase
+    HAVE_BIBTEX_PARSER = True
+except ImportError:
+    HAVE_BIBTEX_PARSER = False
+
+LATEX_OPTS = ['-halt-on-error', '-file-line-error']
 
 if sys.version_info[0] >= 3:
     string_types = (str, bytes)
@@ -142,3 +155,117 @@ def month_and_year(m=None, y=None):
     m = month_to_int(m)
     return '{0} {1}'.format(SHORT_MONTH_NAMES[m], y)
 
+
+def filter_publications(citations, authors, reverse=False, bold=True):
+    pubs = []
+    for pub in citations:
+        if len(set(pub['author']) & authors) == 0:
+            continue
+        pub = deepcopy(pub)
+        if bold:
+            bold_self = []
+            for a in pub['author']:
+                if a in authors:
+                    bold_self.append('\\textbf{' + a + '}')
+                else:
+                    bold_self.append(a)
+            pub['author'] = bold_self
+        else:
+            pub = deepcopy(pub)
+        pubs.append(pub)
+    pubs.sort(key=doc_date_key, reverse=reverse)
+    return pubs
+
+
+def filter_projects(projects, authors, reverse=False):
+    projs = []
+    for proj in projects:
+        team_names = set(gets(proj['team'], 'name'))
+        if len(team_names & authors) == 0:
+            continue
+        proj = dict(proj)
+        proj['team'] = [x for x in proj['team'] if x['name'] in authors]
+        projs.append(proj)
+    projs.sort(key=id_key, reverse=reverse)
+    return projs
+
+
+def filter_grants(input_grants, names, pi=True, reverse=True):
+    grants = []
+    total_amount = 0.0
+    subaward_amount = 0.0
+    for grant in input_grants:
+        team_names = set(gets(grant['team'], 'name'))
+        if len(team_names & names) == 0:
+            continue
+        grant = deepcopy(grant)
+        person = [x for x in grant['team'] if x['name'] in names][0]
+        if pi:
+            if person['position'].lower() == 'pi':
+                total_amount += grant['amount']
+            else:
+                continue
+        else:
+            if person['position'].lower() == 'pi':
+                continue
+            else:
+                total_amount += grant['amount']
+                subaward_amount += person.get('subaward_amount', 0.0)
+                grant['subaward_amount'] = person.get('subaward_amount',
+                                                      0.0)
+                grant['pi'] = [x for x in grant['team'] if
+                               x['position'].lower() == 'pi'][0]
+                grant['me'] = person
+        grants.append(grant)
+    grants.sort(key=ene_date_key, reverse=reverse)
+    return grants, total_amount, subaward_amount
+
+
+def awards_grants_honors(p):
+    """Make sorted awards grants and honors list."""
+    aghs = []
+    for x in p.get('funding', ()):
+        d = {'description': '{0} ({1}{2:,})'.format(
+            latex_safe(x['name']),
+            x.get('currency', '$').replace('$', '\$'), x['value']),
+            'year': x['year'],
+            '_key': date_to_float(x['year'], x.get('month', 0)),
+        }
+        aghs.append(d)
+    for x in p.get('service', []) + p.get('honors', []):
+        d = {'description': latex_safe(x['name']),
+             'year': x['year'],
+             '_key': date_to_float(x['year'], x.get('month', 0)),
+             }
+        aghs.append(d)
+    aghs.sort(key=(lambda x: x.get('_key', 0.0)), reverse=True)
+    return aghs
+
+
+def latex_safe(s):
+    return s.replace('&', '\&').replace('$', '\$').replace('#', '\#')
+
+
+def make_bibtex_file(pubs, pid, person_dir='.'):
+    if not HAVE_BIBTEX_PARSER:
+        return None
+    skip_keys = {'ID', 'ENTRYTYPE', 'author'}
+    bibdb = BibDatabase()
+    bibwriter = BibTexWriter()
+    bibdb.entries = ents = []
+    for pub in pubs:
+        ent = dict(pub)
+        ent['ID'] = ent.pop('_id')
+        ent['ENTRYTYPE'] = ent.pop('entrytype')
+        for n in ['author', 'editor']:
+            if n in ent:
+                ent[n] = ' and '.join(ent[n])
+        for key in ent.keys():
+            if key in skip_keys:
+                continue
+            ent[key] = latex_safe(ent[key])
+        ents.append(ent)
+    fname = os.path.join(person_dir, pid) + '.bib'
+    with open(fname, 'w') as f:
+        f.write(bibwriter.write(bibdb))
+    return fname
